requirements.txt
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
black==24.10.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cloudinary==1.44.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
fastapi-mail==1.6.2
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.4
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
regex==2025.11.3
rich==14.3.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
